Added velocity evolution
Haven't really considered boundary conditions, either at the edges of the grid or within the grid where there is no liquid. Hence the liquid falls off the bottom of the grid when gravity is applied, and generally escapes from the grid.

(I also suspect volume would not be conserved even if we got boundary conditions in place at the edges of the grid.)