Experiment with allowing nonzero velocity away from walls
Applying a more relaxed boundary condition that allows liquid to move away from the boundary walls
didn't work too well. If we apply the looser condition before and after, we get liquid that seems
to be actively repelled from the wall somewhat at random, which is weird. If we apply the zero-velocity
condition before and the looser one after, there's no visible difference from applying zero-velocity
before and after (which I guess makes sense since the pressure conditions we apply are designed to
avoid changing the pre-solve velocity at the boundary).