Avoid negative P at boundaries
Ah, that's better - the looser post-solve velocity boundary condition combined with a tweak to
the pressure boundary condition that contrains P >= 0 at the boundary *does* stop the liquid
sticking to the walls. Most noticeable when you reverse gravity - the liquid no longer wants to
flow around the sides of the box.